backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.28
alembic==1.13.1
pymysql==1.1.0
cryptography==42.0.5
redis==5.0.3
celery==5.3.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
pydantic==2.6.4
pydantic-settings==2.2.1
python-dotenv==1.0.1
python-slugify==8.0.4
aiofiles==23.2.1
gitpython==3.1.42
psutil==5.9.8
